Aplicacao completa. Tela de detalhes completa

+++ app/src/main/java/com/example/leonardo/popularmovies/api/MovieAPIInterface.java
@@ -19,6 +19,6 @@ public interface MovieAPIInterface {
 
     MoviePaginatedResult listPopular(int page, String locale);
 
-    Movie getDetails(int id, String locale);
+    Movie getDetails(long id, String locale);
     String getImageUrl(String imageId, ImageSize imageSize);
 }
